+++ NEWS.md
@@ -1,5 +1,8 @@
 # testthat 1.0.2.9000
 
+* Updated Catch to 1.9.6. `testthat` now understands and makes use of the package
+  routine registration mechanism required by CRAN with R >= 3.4.0.
+
 * `test_examples()` now works with installed packages as well as source
   packages (@jimhester, #532).